split SngFileHeaderValidation and utils #48
